type update + proposition starts with 0 votants

+++ proposition/src/main/java/org/ipi/proposition/model/Proposition.java
@@ -43,10 +43,10 @@ public class Proposition extends PropositionGenerale{
     @CollectionTable(name = "EQUIPES", joinColumns = @JoinColumn(name = "PROPRIETAIRE"))
     @Column(name = "EQUIPE_ID")
     private List<Integer> equipes;
-    @ElementCollection(targetClass=Integer.class)
+    @ElementCollection(targetClass=Long.class)
     @CollectionTable(name = "VOTANTS", joinColumns = @JoinColumn(name = "EQUIPE_ID"))
     @Column(name = "VOTANTS_ID")
-    private List<Integer> votants;
+    private List<Long> votants;
 //    @Column(name = "DATEDEPOT")
 //    private Date dateDepot;
